+++ usr.sbin/smtpd/smtp_session.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: smtp_session.c,v 1.436 2023/10/09 20:55:32 millert Exp $	*/
+/*	$OpenBSD: smtp_session.c,v 1.437 2023/11/03 13:38:28 op Exp $	*/
 
 /*
  * Copyright (c) 2008 Gilles Chehade <gilles@poolp.org>
@@ -2279,7 +2279,7 @@ smtp_auth_failure_pause(struct smtp_session *s)
 	tv.tv_sec = 0;
 	tv.tv_usec = arc4random_uniform(1000000);
 	log_trace(TRACE_SMTP, "smtp: timing-attack protection triggered, "
-	    "will defer answer for %lu microseconds", tv.tv_usec);
+	    "will defer answer for %lu microseconds", (long)tv.tv_usec);
 	evtimer_set(&s->pause, smtp_auth_failure_resume, s);
 	evtimer_add(&s->pause, &tv);
 }
